+++ include/matrix/transformation.h
@@ -3,11 +3,11 @@
 #include "matrix/matrix.h"
 
 namespace rayTracer {
-    matrix translation(const double &x, const double &y, const double &z);
-    matrix scaling(const double &x, const double &y, const double &z);
-    matrix xRotation(const double &r);
-    matrix yRotation(const double &r);
-    matrix zRotation(const double &r);
-    matrix shearing(const double &xY, const double &xZ, const double &yX, const double &yZ, const double &zX, const double &zY);
+    matrix translation(const float &x, const float &y, const float &z);
+    matrix scaling(const float &x, const float &y, const float &z);
+    matrix xRotation(const float &r);
+    matrix yRotation(const float &r);
+    matrix zRotation(const float &r);
+    matrix shearing(const float &xY, const float &xZ, const float &yX, const float &yZ, const float &zX, const float &zY);
     matrix viewTransform(const tuple &from, const tuple &to, const tuple &up);
 }
