+++ include/world/camera.h
@@ -9,23 +9,23 @@ namespace rayTracer {
         private:
             int hSize;
             int vSize;
-            double fov;
-            double halfWidth;
-            double halfHeight;
-            double pixelSize;
+            float fov;
+            float halfWidth;
+            float halfHeight;
+            float pixelSize;
             matrix transform;
             matrix inverseTransform;
 
             void calculatePixelSize();
         public:
-            camera(const int &hSize, const int &vSize, const double &fov, const matrix &transform);
-            camera(const int &hSize, const int &vSize, const double &fov);
+            camera(const int &hSize, const int &vSize, const float &fov, const matrix &transform);
+            camera(const int &hSize, const int &vSize, const float &fov);
 
             const int &getHSize() const;
             const int &getVSize() const;
-            const double &getFov() const;
+            const float &getFov() const;
             const matrix &getTransform() const;
-            const double &getPixelSize() const;
+            const float &getPixelSize() const;
 
             ray getRayForPixel(const int &x, const int &y) const;
             canvas render(const world &w) const;
